requirements.txt
alembic==1.15.2
anyio==4.9.0
arrow==1.3.0
binaryornot==0.4.4
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cookiecutter==1.7.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jinja2-time==0.2.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
poyo==0.5.0
prompt_toolkit==3.0.51
pyasn1==0.4.8
pydantic==1.10.21
Pygments==2.19.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.2
questionary==1.10.0
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
text-unidecode==1.3
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0; sys_platform != "win32"
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
